[KAIR-145] Move nudger information to a new page
